% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluefin_catage.R
\docType{data}
\name{bluefin_catage}
\alias{bluefin_catage}
\title{Bluefin Tuna}
\format{Matrix with years as row names and ages as column names.}
\usage{
bluefin_catage
}
\description{
Bluefin tuna data (catch at age) to demonstrate \code{\link{vpa}}.
}
