% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscod_catage.R
\docType{data}
\name{nscod_catage}
\alias{nscod_catage}
\alias{nscod_natmort}
\alias{nscod_survey}
\title{North Sea Cod}
\format{Matrices with years as row names and ages as column names.}
\usage{
nscod_catage
nscod_natmort
nscod_survey
}
\description{
North Sea cod data (catch at age, natural mortality, survey) to demonstrate
\code{\link{sca}}.
}
