% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca.R
\name{sca}
\alias{sca}
\title{Statistical Catch at Age}
\usage{
sca(par, data, full = FALSE)
}
\arguments{
\item{par}{a vector of parameters named \code{logNa}, \code{logNt},
\code{logFa}, \code{logFt}, and \code{logQ}.}

\item{data}{a list of matrices named \code{C}, \code{I}, and \code{M}.
Alternatively, \code{M} can be a scalar.}

\item{full}{whether to return a full list of objects.}
}
\value{
Objective function value as a scalar, or a full list of objects if
\code{full = TRUE}.
}
\description{
Run simple statistical catch-at-age model: one fleet, one survey, and a plus
group.
}
\details{
The \code{par} vector consists of five blocks:
\tabular{lll}{
  Name         \tab Length           \tab Notes\cr
  \code{logNa} \tab \code{ncol(N}    \tab Age 1 to \eqn{A} in catch\cr
  \code{logNt} \tab \code{nrow(N)-1} \tab Year 2 to \eqn{T+1} in catch\cr
  \code{logFa} \tab \code{ncol(F)}   \tab Age 1 to \eqn{A-1} in catch\cr
  \code{logFt} \tab \code{nrow(F)}   \tab Year 1 to \eqn{T} in catch\cr
  \code{logQ}  \tab \code{ncol(I)}   \tab Age 1 to \eqn{A} in survey
}

Selectivity of the younger ages is relative to the oldest age. The
selectivity of the oldest age is always 1, and is not included in the
parameter vector.
}
\examples{
data <- list(C=nscod_catage, I=nscod_survey, M=nscod_natmort)

logNa <- rep(8, ncol(data$C))
logNt <- rep(8, nrow(data$C))
logFa <- rep(0, ncol(data$C)-1)
logFt <- rep(0, nrow(data$C))
logQ <- rep(-5, ncol(data$I))
par <- c(logNa=logNa, logNt=logNt, logFa=logFa, logFt=logFt, logQ=logQ)

optim(f=sca, par=par, data=data, method="BFGS", control=list(maxit=1000))

}
